Full App
This version allows the user to enter the date of the prices and fixes the positioning of the prices in the table. Furthermore, the app currently doesn't save the image internally and doesn't require any permissions. The app just creates the image and displays it. So the user needs to take a screenshot for the image.